requirements.txt
google-analytics-data
google-auth 
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib 
requests 
xlsxwriter
pandas
instaloader==4.14
celery==5.4.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
gevent==24.10.3
gevent-websocket==0.10.1
google-analytics-data==0.18.15
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
instaloader==4.14
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
kombu==5.4.2
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.1.0
numpy==2.1.2
oauthlib==3.2.2
ohneio==0.8.1
pandas==2.2.3
phylline==0.1.6
phyllo==0.2.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
wsproto==1.2.0
XlsxWriter==3.2.0
zerobouncesdk==1.1.0
zope.event==5.0
zope.interface==7.1.1